Refactor geometric math functions from geo.js to geom.js
