+++ test/index.html
@@ -74,6 +74,7 @@
 
     <script src='spec/geo/extent.js'></script>
     <script src='spec/geo/geo.js'></script>
+    <script src='spec/geo/geom.js'></script>
     <script src='spec/geo/vector.js'></script>
 
     <script src='spec/lib/d3.combobox.js'></script>
